Finished create table

+++ src/common/Tokens.java
@@ -3,7 +3,7 @@ package common;
 public class Tokens {
     public static final String INT = "INT";
     public static final String LONG = "LONG";
-    public static final String FLOAT = "LONG";
+    public static final String FLOAT = "FLOAT";
     public static final String DOUBLE = "DOUBLE";
     public static final String STRING = "STRING";
 }
